API/requirements.txt
fastapi==0.104.1
mlflow==2.8.1
numpy==1.26.2
Pillow==10.1.0
uvicorn==0.24.0.post1
tensorflow==2.14.0
python-multipart==0.0.6
